starting work on transports and bus (wip)
+ converted weakmaps to symbols in service locator.  didn't need absolute certainly on private variables so symbols are fine.
+ more work on PbjxEvent to establish mirror of php features

first transports will be in memory (for tests) and "fetch" for single page apps.

The fetch transport is designed with the assumption that no "handling" can be done by the app calling the pbjx.send/request and it must hit a server.  This is similar to using "Firehose" transport in php because it's assuming that the "sender" can't handle the command if the transport service is down and is likely not privileged enough to do so anyways.