small code optimizations for tokens, also use standard jwt claims
changing our use of "host" and "pbjx" to use "aud" and "jti" as they are basically what we're using the jwt spec as.  The host aka aud (audience) claim tells us what the intended server endpoint this token is for and is exactly what we want.

The "jti" is a nonce which is also what we are looking for with the pbjx (a hash of the content) which should only be used once.  We combine the aud and iat with that content to further ensure its uniqueness as it's quite possible to have the same payload but not nearly as likely to have the same at exactly the same time for the same aud (endpoint).

By using the reserved jwt claims we might be able to generalize the validation or use other tooling to create/validate the jwt at different layers of the stack.